requirements.txt
streamlit>=1.38.0
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
sqlalchemy>=2.0.0
aiosqlite>=0.19.0
vnstock>=3.1.0
httpx>=0.27.0
aiohttp>=3.9.0
websockets>=12.0
python-binance>=1.0.19
requests>=2.31.0
google-generativeai>=0.8.0
openai>=1.40.0
langchain>=0.3.0
langchain-google-genai>=2.0.0
scikit-learn>=1.4.0
xgboost>=2.0.0
pandas>=2.2.0
numpy>=1.26.0
pandas-ta
plotly>=5.20.0
matplotlib>=3.8.0
seaborn>=0.13.0
mplfinance>=0.12.10b0
python-telegram-bot>=21.0
loguru>=0.7.2
apscheduler>=3.10.4
bcrypt>=4.1.0
pyjwt>=2.8.0
openpyxl>=3.1.0
xlsxwriter>=3.2.0
nest-asyncio>=1.6.0
# vnstock ecosystem (cài qua vnstock-installer nếu là gói Insiders)
vnstock_data
vnstock_ta
vnstock_news
vnstock_chart
vnstock_pipeline
